Unify all size_t arguments to use uint
Somewhat addresses #212
* Updated to use uint consistently where CSFML is expecting size_t
* UIntPtr is the most correct type for size_t, but switching to it would result in breaking changes for end users
* uint should be replaced with nuint when moving to .NET 5+

+++ src/SFML.Audio/Music.cs
@@ -390,7 +390,7 @@ namespace SFML.Audio
         private unsafe static extern IntPtr sfMusic_createFromStream(IntPtr stream);
 
         [DllImport(CSFML.audio, CallingConvention = CallingConvention.Cdecl), SuppressUnmanagedCodeSecurity]
-        private static extern IntPtr sfMusic_createFromMemory(IntPtr data, ulong size);
+        private static extern IntPtr sfMusic_createFromMemory(IntPtr data, uint size);
 
         [DllImport(CSFML.audio, CallingConvention = CallingConvention.Cdecl), SuppressUnmanagedCodeSecurity]
         private static extern void sfMusic_destroy(IntPtr MusicStream);
